app5/requirements.txt
autogen-agentchat[llm,gemini]~=0.2
agentops